# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "UReport API",
    "version": "2.0.0",
    "description": "UReport REST API"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Log in with username and password",
        "description": "",
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Account not active"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        }
      }
    },
...
    "/user/forgot": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Request a password reset email",
        "description": "",
        "responses": {
          "200": {
            "description": "Reset email sent if account exists"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/build": {
      "post": {
        "tags": [
          "Build"
        ],
        "summary": "Create or update a build",
        "description": "",
        "responses": {
          "200": {
            "description": "Build created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Build"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Build"
              }
            }
          }
        }
      }
    },
...
    "/test_relation": {
      "post": {
        "tags": [
          "TestRelation"
        ],
        "summary": "Create or update a test relation",
        "description": "",
        "responses": {
          "200": {
            "description": "TestRelation created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRelation"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestRelation"
              }
            }
          }
        }
      }
    },
...
    "/investigated_test": {
      "post": {
        "tags": [
          "InvestigatedTest"
        ],
        "summary": "Create or update an investigated test",
        "description": "",
        "responses": {
          "200": {
            "description": "InvestigatedTest created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigatedTest"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigatedTest"
              }
            }
          }
        }
      }
    },
...
    "/assignment": {
      "post": {
        "tags": [
          "Assignment"
        ],
        "summary": "Create or update an assignment",
        "description": "",
        "responses": {
          "200": {
            "description": "Assignment created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assignment"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Assignment"
              }
            }
          }
        }
      }
    },
...
    "/dashboard": {
      "post": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Create or update a dashboard",
        "description": "",
        "responses": {
          "200": {
            "description": "Dashboard created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Dashboard"
              }
            }
          }
        }
      }
    },
...
    "/template": {
      "post": {
        "tags": [
          "DashboardTemplate"
        ],
        "summary": "Create or update a dashboard template",
        "description": "",
        "responses": {
          "200": {
            "description": "Template created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardTemplate"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardTemplate"
              }
            }
          }
        }
      }
    },
...
    "/setting": {
      "post": {
        "tags": [
          "Setting"
        ],
        "summary": "Create a new setting",
        "description": "",
        "responses": {
          "200": {
            "description": "Setting created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Setting"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Setting"
              }
            }
          }
        }
      }
    },
...
    "/system/setting": {
      "post": {
        "tags": [
          "SystemSetting"
        ],
        "summary": "Create a system setting",
        "description": "",
        "responses": {
          "200": {
            "description": "System setting created"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/preset": {
      "get": {
        "tags": [
          "Preset"
        ],
        "summary": "Get all presets",
        "description": "",
        "responses": {
          "200": {
            "description": "List of presets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Preset"
        ],
        "summary": "Create a preset",
        "description": "",
        "responses": {
          "201": {
            "description": "Preset created"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "lanes"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "lanes": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/user/favorite": {
      "post": {
        "tags": [
          "UserFavorite"
        ],
        "summary": "Create or update user favorites",
        "description": "",
        "responses": {
          "200": {
            "description": "Saved favorites"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userid"
                ],
                "properties": {
                  "userid": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "session": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          }
        }
      },
      "Build": {
        "type": "object",
        "required": [
          "product",
          "type",
          "build"
        ],
        "properties": {
          "product": {
            "type": "string",
            "example": "Reporting",
            "description": "The product name of the software."
          },
          "type": {
            "type": "string",
            "example": "API or UI or Regression or Smoke",
            "description": "The type of the automation."
          },
          "build": {
            "type": "integer",
            "example": 100,
            "description": "Build number of the execution."
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          },
          "team": {
            "type": "string",
            "example": "Team Alpha"
          },
          "stage": {
            "type": "string",
            "example": "staging",
            "description": "Stage of the automation run (e.g. staging, production)."
          },
          "browser": {
            "type": "string",
            "enum": [
              "Chrome",
              "Firefox",
              "Edge",
              "Safari",
              "Opera",
              "IE",
              "Internet Explorer",
              "Electron",
              "chromium"
            ],
            "example": "Chrome"
          },
          "device": {
            "type": "string",
            "example": "Windows VM or Pixel 2"
          },
          "platform": {
            "type": "string",
            "example": "iOS or Windows"
          },
          "platform_version": {
            "type": "string",
            "example": "13.0.4 or 10"
          },
          "start_time": {
            "type": "string",
            "format": "date-time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time"
          },
          "owner": {
            "type": "string",
            "example": "admin"
          },
          "is_archive": {
            "type": "boolean",
            "default": false
          },
          "status": {
            "type": "object",
            "properties": {
              "pass": {
                "type": "integer",
                "default": 0
              },
              "fail": {
                "type": "integer",
                "default": 0
              },
              "skip": {
                "type": "integer",
                "default": 0
              },
              "warning": {
                "type": "integer",
                "default": 0
              },
              "total": {
                "type": "integer",
                "default": 0
              }
            }
          },
          "environments": {
            "type": "object",
            "description": "Environment variables of the automation run."
          },
          "settings": {
            "type": "object",
            "description": "Settings variables of the automation run."
          },
          "client": {
            "type": "string",
            "example": "VM-01",
            "description": "If provided, settings/environments are saved under the client name."
          }
        }
      },
...
      "TestRelation": {
        "type": "object",
        "required": [
          "uid",
          "product",
          "type"
        ],
        "properties": {
          "uid": {
            "type": "string"
          },
          "product": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "file": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "components": {
            "type": "object"
          },
          "teams": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Flexible tag objects (Mixed type)."
          },
          "customs": {
            "type": "object",
            "description": "Any custom key-value pairs (Mixed type)."
          }
        }
      },
      "InvestigatedTest": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string"
          },
          "product": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "user": {
            "type": "string"
          },
          "caused_by": {
            "type": "string",
            "default": "Defect",
            "description": "Root cause category."
          },
          "tracking": {
            "type": "string",
            "description": "Tracking ticket reference."
          },
          "failure": {
            "type": "object"
          },
          "origin": {
            "type": "string"
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "configuration": {
            "type": "object"
          },
          "customize_state": {
            "type": "string"
          }
        }
      },
      "Assignment": {
        "type": "object",
        "properties": {
          "product": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "user": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "uid": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "default": "OPEN"
          },
          "failure": {
            "type": "object"
          },
          "test_url": {
            "type": "string"
          },
          "assign_at": {
            "type": "string",
            "format": "date-time"
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "Setting": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "object"
          }
        }
      },
      "Dashboard": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "product": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "widgets": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "DashboardTemplate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "widgets": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "is_public": {
            "type": "boolean",
            "default": false
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}